+++ src/CommandCore/PolymorphicReversibleCommand.h
@@ -51,11 +51,7 @@ private:
     template<class Archive>
     void serialize(Archive& archive)
     {
-#if COOL_SERIALIZATION
         archive(_concrete_reversible_command);
-#else
-        (void)archive;
-#endif
     }
 };