+++ src/Commands/Command_FinishedEditingVariable.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "CommandCore/CommandExecutionContext_Ref.h"
-#include "CommandCore/LAB_REGISTER_COMMAND.h"
 
 namespace Lab {
 
@@ -30,5 +29,3 @@ void serialize(Archive&, Lab::Command_FinishedEditingVariable&)
 }
 
 } // namespace cereal
-
-LAB_REGISTER_COMMAND(Lab::Command_FinishedEditingVariable)