+++ src/Commands/Command_SetCameraZoom.cpp
@@ -1,0 +1,6 @@
+#include "Command_SetCameraZoom.h"
+#include "CommandCore/LAB_REGISTER_COMMAND.h"
+#include "CommandCore/LAB_REGISTER_REVERSIBLE_COMMAND.h"
+
+LAB_REGISTER_COMMAND(Lab::Command_SetCameraZoom)
+LAB_REGISTER_REVERSIBLE_COMMAND(Lab::ReversibleCommand_SetCameraZoom)