+++ src/Commands/Command_SetVariableMetadata.h
@@ -2,7 +2,6 @@
 
 #include <stringify/stringify.hpp>
 #include "CommandCore/CommandExecutionContext_Ref.h"
-#include "CommandCore/LAB_REGISTER_COMMAND.h"
 
 namespace Lab {
 
@@ -31,17 +30,10 @@ namespace cereal {
 template<class Archive, typename T>
 void serialize(Archive& archive, Lab::Command_SetVariableMetadata<T>& command)
 {
-#if COOL_SERIALIZATION
     archive(
         cereal::make_nvp("Id", command.id),
         cereal::make_nvp("Metadata", command.metadata)
     );
-#else
-    (void)archive;
-    (void)command;
-#endif
 }
 
 } // namespace cereal
-
-#include "generated/register_set_variable_metadata_commands.inl"