+++ src/Commands/generate_commands.py
@@ -26,8 +26,6 @@ def command_set(cmd: SetterCommand):
     return f"""
 #include <stringify/stringify.hpp>
 #include "CommandCore/CommandExecutionContext_Ref.h"
-#include "CommandCore/LAB_REGISTER_COMMAND.h"
-#include "CommandCore/LAB_REGISTER_REVERSIBLE_COMMAND.h"
 #include "CommandCore/MakeReversibleCommandContext_Ref.h"
 
 namespace Lab {{
@@ -95,32 +93,20 @@ namespace cereal {{
 template<class Archive>
 void serialize(Archive& archive, Lab::Command_{cmd.name}& command)
 {{
-#if COOL_SERIALIZATION
     archive(cereal::make_nvp("{cmd.user_facing_name}", command.value));
-#else
-    (void)archive;
-    (void)command;
-#endif
 }}
 
 template<class Archive>
 void serialize(Archive& archive, Lab::ReversibleCommand_{cmd.name}& command)
 {{
-#if COOL_SERIALIZATION
     archive(
         cereal::make_nvp("{cmd.user_facing_name}", command.forward_command.value),
         cereal::make_nvp("Old {cmd.user_facing_name}", command.old_value)
     );
-#else
-    (void)archive;
-    (void)command;
-#endif
 }}
 
 }} // namespace cereal
 
-LAB_REGISTER_COMMAND(Lab::Command_{cmd.name})
-LAB_REGISTER_REVERSIBLE_COMMAND(Lab::ReversibleCommand_{cmd.name})
 """