+++ src/Debug/DebugOptions.cpp
@@ -1,0 +1,2 @@
+#include "DebugOptions.h"
+#include "generated/DebugOptionsCpp.inl"