+++ src/Dependencies/Camera2DManager.cpp
@@ -17,12 +17,16 @@ void hook_camera2D_events(
     Cool::Camera2D&                                    camera,
     std::function<void()>                              on_change,
     std::function<float()>                             get_height,
-    std::function<float()>                             get_aspect_ratio
+    std::function<float()>                             get_aspect_ratio,
+    std::function<bool()>                              is_locked_in_view
 )
 {
     events
         .scroll_event()
-        .subscribe([&, on_change, get_height, get_aspect_ratio](Cool::MouseScrollEvent<Cool::ViewCoordinates> const& event) {
+        .subscribe([&, on_change, get_height, get_aspect_ratio, is_locked_in_view](Cool::MouseScrollEvent<Cool::ViewCoordinates> const& event) {
+            if (is_locked_in_view())
+                return;
+
             float const sensitivity    = Cool::user_settings().camera2D_zoom_sensitivity;
             float const zoom_variation = std::pow(sensitivity, event.dy);
 
@@ -38,7 +42,10 @@ void hook_camera2D_events(
     events
         .drag()
         .update()
-        .subscribe([&, on_change, get_height](Cool::MouseDragUpdateEvent<Cool::ViewCoordinates> const& event) {
+        .subscribe([&, on_change, get_height, is_locked_in_view](Cool::MouseDragUpdateEvent<Cool::ViewCoordinates> const& event) {
+            if (is_locked_in_view())
+                return;
+
             camera.translation -= event.delta / get_height() * 2.f / camera.zoom;
             on_change();
         });