+++ src/Module_Nodes/Node.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <Cool/Dependencies/Input.h>
 #include <Cool/Nodes/Pin.h>
+#include <cereal/types/optional.hpp>
 #include <optional>
 #include <string_view>
 #include "Cool/Nodes/NodeDefinitionIdentifier.h"
@@ -26,24 +27,24 @@ public:
 
     auto id_names() const -> Cool::NodeDefinitionIdentifier { return _id_names; }
 
-    auto input_pins() const -> auto const& { return _input_pins; }
-    auto input_pins() -> auto& { return _input_pins; }
-    auto output_pins() const -> auto const& { return _output_pins; }
-    auto output_pins() -> auto& { return _output_pins; }
-    auto value_inputs() -> auto& { return _value_inputs; }
-    auto value_inputs() const -> auto const& { return _value_inputs; }
+    [[nodiscard]] auto input_pins() const -> auto const& { return _input_pins; }
+    [[nodiscard]] auto input_pins() -> auto& { return _input_pins; }
+    [[nodiscard]] auto output_pins() const -> auto const& { return _output_pins; }
+    [[nodiscard]] auto output_pins() -> auto& { return _output_pins; }
+    [[nodiscard]] auto value_inputs() -> auto& { return _value_inputs; }
+    [[nodiscard]] auto value_inputs() const -> auto const& { return _value_inputs; }
 
     auto number_of_main_input_pins() const -> size_t { return _number_of_main_input_pins; }
     auto main_input_pin(size_t main_input_index) const -> Cool::InputPin const&
     {
         assert(main_input_index < number_of_main_input_pins());
-        return _input_pins[main_input_index];
+        return input_pins()[main_input_index];
     }
-    auto main_output_pin() const -> Cool::OutputPin const& { return _output_pins[0]; }
+    auto main_output_pin() const -> Cool::OutputPin const& { return output_pins()[0]; }
     /// This corresponds to a subset of all the input_pints(); the ones that correspond to an INPUT of the node.
-    auto pin_of_function_input(size_t input_index) const -> Cool::InputPin const& { return _input_pins[_number_of_main_input_pins + input_index]; }
-    auto pin_of_value_input(size_t property_index) const -> Cool::InputPin const& { return _input_pins[_number_of_main_input_pins + _number_of_function_inputs + property_index]; }
-    auto pin_of_output_index(size_t output_index_index) const -> Cool::OutputPin const& { return _output_pins[_number_of_main_input_pins + output_index_index]; }
+    auto pin_of_function_input(size_t input_index) const -> Cool::InputPin const& { return input_pins()[_number_of_main_input_pins + input_index]; }
+    auto pin_of_value_input(size_t property_index) const -> Cool::InputPin const& { return input_pins()[_number_of_main_input_pins + _number_of_function_inputs + property_index]; }
+    auto pin_of_output_index(size_t output_index_index) const -> Cool::OutputPin const& { return output_pins()[_number_of_main_input_pins + output_index_index]; }
 
     /// Only call this if this node is a template node
     auto chosen_any_type() const -> PrimitiveType { return _chosen_any_type.value(); }
@@ -53,11 +54,11 @@ private:
     Cool::NodeDefinitionIdentifier _id_names;
     std::string                    _name{};
 
-    std::vector<Cool::InputPin>  _input_pins;
-    std::vector<Cool::OutputPin> _output_pins;
+    std::vector<Cool::InputPin>  _input_pins{};
+    std::vector<Cool::OutputPin> _output_pins{};
     std::vector<Cool::AnyInput>  _value_inputs;
-    size_t                       _number_of_main_input_pins;
-    size_t                       _number_of_function_inputs;
+    size_t                       _number_of_main_input_pins{};
+    size_t                       _number_of_function_inputs{};
 
     std::optional<PrimitiveType> _chosen_any_type{}; // Only present if the node has `Any` in its signature.
 
@@ -66,7 +67,6 @@ private:
     template<class Archive>
     void serialize(Archive& archive)
     {
-#if COOL_SERIALIZATION
         archive(
             cereal::make_nvp("Name", _name),
             cereal::make_nvp("Definition", _id_names),
@@ -77,9 +77,6 @@ private:
             cereal::make_nvp("Number of function inputs", _number_of_function_inputs),
             cereal::make_nvp("Chosen Any type", _chosen_any_type)
         );
-#else
-        (void)archive;
-#endif
     }
 };