+++ changelog.md
@@ -12,8 +12,12 @@
 - 💄 Added a Light color theme.
 - 💄 Added an option to use the color theme set by your OS (Dark or Light).
 - 💄 Moved "About" and "Debug" menus to the right of the menu bar.
+- 💄 The View now uses a fixed aspect ratio by default (you can change this in the Preview menu).
+- 💄 Added a toggle widget that replaces the checkboxes.
+- 💄 Improved Cameras window + option to lock one of the two cameras (2D or 3D) when using both 2D and 3D nodes.
 - 🐛 Fixed the camera controls when using a fixed aspect ratio or fixed number of pixels.
 - 🐛 Fix: the nodes categories were not sorted on MacOS.
+- 🐛 Fix: could not place a node that had the same name as its category.
 - 👩‍💻 Node files: now support the `#define` macro, just like any glsl file.
 - 👩‍💻 Node files: added boolean and matrices types for INPUTs, main function and helper functions (they can be used anywhere like any other type now).